refactor(FR-2300): move checkSslMismatch to configRules (#5971)
Resolves #5966 (FR-2300)

## Summary
- Move `checkSslMismatch` from `endpointRules.ts` to `configRules.ts` to better reflect its category as a config-level check
- Rename diagnostic id from `ssl-mismatch` to `config-ssl-mismatch` for consistency with the config category
- Update the diagnostic result category from `'endpoint'` to `'config'`
- Move corresponding tests from `endpointRules.test.ts` to `configRules.test.ts`
- Update import in `useWebServerConfigDiagnostics.ts` to reference the new location

## Changed Files
- `react/src/diagnostics/rules/configRules.ts` — added `checkSslMismatch` function
- `react/src/diagnostics/rules/endpointRules.ts` — removed `checkSslMismatch` function
- `react/src/diagnostics/rules/__tests__/configRules.test.ts` — added SSL mismatch tests
- `react/src/diagnostics/rules/__tests__/endpointRules.test.ts` — removed SSL mismatch tests
- `react/src/hooks/useWebServerConfigDiagnostics.ts` — updated import path
