+++ react/src/hooks/useWebServerConfigDiagnostics.ts
@@ -11,10 +11,10 @@ import {
   checkPlaceholderValues,
   checkPluginConfiguration,
   checkResourceLimits,
+  checkSslMismatch,
   checkUrlFields,
   isPlaceholder,
 } from '../diagnostics/rules/configRules';
-import { checkSslMismatch } from '../diagnostics/rules/endpointRules';
 import type { DiagnosticResult } from '../types/diagnostics';
 import { useProxyUrl, useRawConfig } from './useWebUIConfig';
 import { VALID_MENU_KEYS } from './useWebUIMenuItems';
@@ -48,7 +48,6 @@ export function useWebServerConfigDiagnostics(): DiagnosticResult[] {
     if (sslCheck) {
       results.push({
         ...sslCheck,
-        id: 'config-ssl-mismatch',
         category: 'config',
         titleKey: 'diagnostics.ConfigSslMismatch',
         descriptionKey: 'diagnostics.ConfigSslMismatchDesc',
